Added filtering and testing for CSV queue entries
Added filtering and testing for CSV queue entries

Implemented the use of expressions for filtering CSV queue entries by adding the `System.Linq.Expressions` namespace and updating the `ICsvQueueBl` and `ICsvQueueDal` interfaces with a new `Read` method. The `CsvQueueBl` and `CsvQueueDal` classes were updated to implement this method. Unit tests for the new method were added in `CsvQueueBlTests` and `CsvQueueDalTests` classes.

A new `CsvQueueDto` class was added to represent a data access object for CSV queue. The `QueueController` class was added to handle HTTP requests related to the CSV queue, with integration tests included in the new `QueueControllerTests` class. The `CardControllerTests` class was updated with a new integration test for the `QueueFromCsv` method.

Project and package references were updated in `TrelloTenderManager.WebApi.IntegrationTests.csproj` and `TrelloTenderManager.WebApi.csproj` files. The `TrelloTenderManager.sln` file was updated to reflect these changes in the project structure. New HTTP requests for testing the `QueueController` were added to the `TrelloTenderManager.WebApi.http` file.